Creates /src/services folder.
This commit creates the services folder for the project. This is
where all the bussiness logic for the project will go. The services
will be used in the routes. This commit creates the transaction
service witch handles the CRUD opertations for the transaction
model and any other bussiness logic associated with that.