Creates /src/routes/ folder.
This commit creates the routes folder. This folder will house
all the routes for the project. The routes should have NO
bussiness logic in them. Any bussines logic should go into
the services for the routes. This commit creates the routes
for the transaction model using the transaction service. The
CRUD operations are currently working for transactions.